requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
idna==3.4
optional-django==0.3.0
psycopg2==2.9.9
python-dotenv==1.0.0
react==4.3.0
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
